// src/app/admin/interfaces/capacitador.interface.ts
// Generated by https://quicktype.io

import { AreaConocimiento } from "./areasConocimiento.interfaace";

// Generated by https://quicktype.io

export interface Capacitador {
  id:              number;
  nombreDocente:   string;
  apellidoDocente: string;
  identificacion:  string;
  correo:          string;
  carrera:         string;
  telefono:        string;
  imagen:          string;
  descripcion:     string;
  estado:          string;
  areas:           AreaConocimiento[];
}